+++ config.js
@@ -41,6 +41,7 @@ export default {
   buildTileUrlTemplate: null,
   getMapSourceOptions: null,
   pathPrefix: "/",
+  runtimeConfig: false,
   historyMode: "history",
   cardViewMode: "cards",
   defaultCollectionSort: "title",